+++ SequencableUI.cpp
@@ -1,0 +1,10 @@
+/*
+    SequencableUI.cpp - Library for interfacing the sequencer with UIViews.
+    Created by Daniel Esteban, February 5, 2013.
+*/
+
+#include "SequencableUI.h"
+
+SequencableUI::SequencableUI() {
+	
+}